Fix data type

+++ shelly-emulator/src/main/java/com/github/n4zroth/sungather/shellyemulator/model/ShellyMessageTotal.java
@@ -3,6 +3,6 @@ package com.github.n4zroth.sungather.shellyemulator.model;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 // emdata:0
-public record ShellyMessageTotal(@JsonProperty("total_act") long totalPower) {
+public record ShellyMessageTotal(@JsonProperty("total_act") Double totalPower) {
 
 }
